Separate test cases from upstream
